Fix : Change in styled file
